Sign the pre-release APK, too
Because it makes it easier for end users to install a beta build if they
want to, meaning more potential eyeballs on upcoming releases.

It is still a debug APK, but a signed one. This probably goes against
Android conventions, but I don't see why it should be a problem.