Git sync: Show branch information in namesake status
A slight improvement: Only alert the user when a *new* merge conflict
occurs. The rest of the time, the sync status message shows the user if
they are not on the main branch, and thus needs to resolve a merge
conflict. They can also see clearly when that conflict has been
successfully resolved.