WIP IntegrallySyncedRepos
Note that I removed the push during storeBook(), because it meant
multiple pushes during a single sync event.

This is potentially problematic, since storeBook() is called during
force-saving of a book, and in such a scenario the user might expect the
repo to be synced. I don't know how to solve this. We have the same
problem with retrieveBook().

I believe the best option would be to add explicit forceSaveBook() and
forceLoadBook() methods to the SyncRepo interface, to allow for greater
flexibility.