Don't ensure our repo is up to date during retrieveBook()
This is hard. retrieveBook() is called both during regular syncing, and
during force-loading of a book. The same goes for storeBook().

We don't want to touch the general logic, but we also don't want to
fetch and push more than once during a sync session. I don't know how to
solve this.

My current "solution" is to require the user to initiate a sync if they
want to make sure that the Git repo is synced with remote. Force-saving
and force-loading will not sync with the remote side, in itself. I know
this is bad, but it's the least bad option I can see right now.