Fix bugs in handling of non-Git repos
I haven't found a decent way to ignore books and vrooks in
IntegrallySyncedRepos when syncing the other types of repos.

There is the RepoType enumeration, but those are just the concrete
implementations, and I want to identify IntegrallySyncedRepos vs other
SyncRepos. I don't really see how I can do that when going through the
local books from SyncWorker or SyncUtils. The local books can have a
linkedRepo, which has a RepoType, but that's all I can see. There is no
connection between RepoType.GIT and the IntegrallySyncedRepo interface.