Don't instantiate external access action handlers upon app launch
Since external access typically happens either infrequently or in an
automated fashion, any small delay this instantiation creates should be
tolerable.

The reason for this change is that it allows us to run local JVM tests.
Before this change, the app could not run without a real Android device
and its Context.