Clean up handling of the SCHEDULE_EXACT_ALARM permission
Instead of trying to make the same check in several places (and fail at
it) we'll just show a snackbar whenever we fail to schedule reminders,
which the user can click to reach the relevant app settings screen.

Also, stop using AlarmManager.setExact() for daily reminders, as that
method requires the "schedule exact alarm" permission for no good
reason. Use AlarmManager.set() instead, which means that the
notification will be delivered "approximately" at the specified time.

This should fix issue #473.